# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Solennix API
  version: 1.0.0
  description: Event management platform for Latin American organizers
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://api.solennix.com
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      required:
        - id
        - user_id
        - client_id
        - event_date
        - service_type
        - num_people
        - status
        - discount
        - discount_type
        - requires_invoice
        - tax_rate
        - tax_amount
        - total_amount
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        event_date:
          type: string
          description: DATE string in YYYY-MM-DD format
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        service_type:
          type: string
        num_people:
          type: integer
        status:
          type: string
          enum: [quoted, confirmed, completed, cancelled]
        discount:
          type: number
          format: double
        discount_type:
          type: string
          enum: [percent, fixed]
        requires_invoice:
          type: boolean
        tax_rate:
          type: number
          format: double
        tax_amount:
          type: number
          format: double
        total_amount:
          type: number
          format: double
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        deposit_percent:
          type: number
          format: double
          nullable: true
        cancellation_days:
          type: number
          format: double
          nullable: true
        refund_percent:
          type: number
          format: double
          nullable: true
        notes:
          type: string
          nullable: true
        photos:
          type: string
          nullable: true
        client:
          $ref: "#/components/schemas/Client"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EventUpsertRequest:
      type: object
      required:
        - client_id
        - event_date
        - service_type
        - num_people
        - status
        - discount
        - discount_type
        - requires_invoice
        - tax_rate
        - tax_amount
        - total_amount
      properties:
        client_id:
          type: string
          format: uuid
        event_date:
          type: string
          description: DATE string in YYYY-MM-DD or ISO-like value; backend validates date portion
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        service_type:
          type: string
        num_people:
          type: integer
        status:
          type: string
          enum: [quoted, confirmed, completed, cancelled]
        discount:
          type: number
          format: double
        discount_type:
          type: string
          enum: [percent, fixed]
        requires_invoice:
          type: boolean
        tax_rate:
          type: number
          format: double
        tax_amount:
          type: number
          format: double
        total_amount:
          type: number
          format: double
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        deposit_percent:
          type: number
          format: double
          nullable: true
        cancellation_days:
          type: number
          format: double
          nullable: true
        refund_percent:
          type: number
          format: double
          nullable: true
        notes:
          type: string
          nullable: true
    Client:
      type: object
      required:
        - id
        - user_id
        - name
        - phone
        - total_events
        - total_spent
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        total_events:
          type: integer
        total_spent:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginatedEventsResponse:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - total_pages
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code or type
        message:
          type: string
          description: Human-readable error message
...
    EventPublicLink:
      type: object
      required:
        - id
        - event_id
        - user_id
        - token
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        token:
          type: string
          description: 64-character hex token
        status:
          type: string
          enum: [active, revoked, expired]
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        url:
          type: string
          description: Computed client-portal URL (not stored in DB)
    EventPublicLinkCreateRequest:
      type: object
      properties:
        ttl_days:
          type: integer
          minimum: 1
          maximum: 730
          nullable: true
          description: Link validity in days (1–730, up to 2 years). Omit or null for no expiry. Higher limit than EventFormLink because client-portal links are long-lived.
    PublicEventDetails:
      type: object
      required:
        - id
        - service_type
        - event_date
        - num_people
        - status
      properties:
        id:
          type: string
          format: uuid
        service_type:
          type: string
        event_date:
          type: string
          description: Date in YYYY-MM-DD format
        start_time:
          type: string
          nullable: true
        end_time:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        num_people:
          type: integer
        status:
          type: string
          enum: [quoted, confirmed, completed, cancelled]
    PublicOrganizerBrand:
      type: object
      properties:
        business_name:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        brand_color:
          type: string
          nullable: true
    PublicClientInfo:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    PublicPaymentSummary:
      type: object
      required:
        - total
        - paid
        - remaining
        - currency
      properties:
        total:
          type: number
          format: double
        paid:
          type: number
          format: double
        remaining:
          type: number
          format: double
        currency:
          type: string
          description: ISO 4217 currency code (e.g. MXN)
    PublicEventView:
      type: object
      required:
        - portal_tier
        - capabilities
        - event
        - organizer
        - client
        - payment
      properties:
        portal_tier:
          type: string
          description: Portal tier behavior for this tokenized view
          enum: [free, pro]
        capabilities:
          $ref: "#/components/schemas/PublicPortalCapabilities"
        event:
          $ref: "#/components/schemas/PublicEventDetails"
        organizer:
          $ref: "#/components/schemas/PublicOrganizerBrand"
        client:
          $ref: "#/components/schemas/PublicClientInfo"
        payment:
          $ref: "#/components/schemas/PublicPaymentSummary"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/PublicPortalMilestone"
    PublicPortalCapabilities:
      type: object
      required:
        - can_submit_payment
        - can_view_submission_history
        - can_view_rich_event_details
        - can_view_milestones
        - can_download_receipt_files
      properties:
        can_submit_payment:
          type: boolean
        can_view_submission_history:
          type: boolean
        can_view_rich_event_details:
          type: boolean
        can_view_milestones:
          type: boolean
        can_download_receipt_files:
          type: boolean
    PublicPortalMilestone:
      type: object
      required:
        - label
      properties:
        label:
          type: string
        date:
          type: string
        status:
          type: string
        highlight:
          type: boolean
...
paths:
...
  /api/events:
    get:
      operationId: listEvents
      summary: Get all events for authenticated user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of events or paginated events when page is provided
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  - $ref: "#/components/schemas/PaginatedEventsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createEvent
      summary: Create new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpsertRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/events/{id}:
    get:
      operationId: getEvent
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
        "404":
          description: Event not found
    put:
      operationId: updateEvent
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpsertRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Event not found
    delete:
      operationId: deleteEvent
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Event deleted
        "401":
          description: Unauthorized
        "404":
          description: Event not found
...
  /api/events/{id}/public-link:
    get:
      operationId: getEventPublicLink
      summary: Get the currently active client-portal link for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Active public link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPublicLink"
        "400":
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: No active public link for this event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get public link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrRotateEventPublicLink
      summary: Create or rotate the client-portal link for an event
      description: |
        Creates a new active public link, revoking any previously active one.
        Repeated calls rotate the token, invalidating any URL already shared.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventPublicLinkCreateRequest"
      responses:
        "201":
          description: Public link created or rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPublicLink"
        "400":
          description: Invalid event ID or ttl_days out of range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to create public link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: revokeEventPublicLink
      summary: Revoke the active client-portal link for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Public link revoked
        "400":
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "404":
          description: No active public link to revoke
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to revoke public link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/public/events/{token}:
    get:
      security: []
      operationId: getEventPortalData
      summary: Get read-only client portal data for a shared event link
      description: |
        Public endpoint (no auth). Rate-limited to 10 requests/minute.
        The token is the proof of access — no authentication required.
        Responds 404 for unknown tokens and 410 Gone for revoked or expired links.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client portal data with event details, organizer branding, and payment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEventView"
        "400":
          description: Token is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Link has been revoked, expired, or the associated event was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to load portal data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...